sala-debate/backend/requirements.txt
agentscope
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
python-dotenv==1.0.1
eventlet==0.33.3
psycopg2-binary==2.9.6
google-generativeai
eventlet